refactoring

+++ tokenizer.c
@@ -12,7 +12,8 @@ LVar *locals;
 // 変数を名前で検索する。見つからなかった場合はNULLを返す。
 LVar *find_lvar(Token *tok) {
   for (LVar *var = locals; var; var = var->next)
-    if (var->len == tok->len && !memcmp(tok->str, var->name, var->len))
+    if (var->len == tok->len &&
+        !memcmp(tok->str, var->name, (unsigned int)var->len))
       return var;
   return NULL;
 }
@@ -140,14 +141,14 @@ Token *tokenize(char *p) {
       } while (is_alnum(*p) || is_digit(*p));
 
       cur = new_token(TK_IDENT, cur, start);
-      cur->len = p - start;
+      cur->len = (int)(p - start);
       continue;
     }
 
     // 数字
     if (is_digit(*p)) {
       cur = new_token(TK_NUM, cur, p);
-      cur->val = strtol(p, &p, 10);
+      cur->val = (int)strtol(p, &p, 10);
       continue;
     }
 
@@ -165,8 +166,8 @@ bool consume_reserved(Token **token, TokenKind kind) {
 }
 
 bool consume(Token **token, char *op) {
-  if ((*token)->kind != TK_RESERVED || strlen(op) != (*token)->len ||
-      memcmp((*token)->str, op, (*token)->len))
+  if ((*token)->kind != TK_RESERVED || (int)strlen(op) != (*token)->len ||
+      memcmp((*token)->str, op, (unsigned long)(*token)->len))
     return false;
   (*token) = (*token)->next;
   return true;
@@ -180,8 +181,8 @@ Token *consume_ident(Token **token) {
 }
 
 void expect(Token **token, char *op) {
-  if ((*token)->kind != TK_RESERVED || strlen(op) != (*token)->len ||
-      memcmp((*token)->str, op, (*token)->len))
+  if ((*token)->kind != TK_RESERVED || (int)strlen(op) != (*token)->len ||
+      memcmp((*token)->str, op, (unsigned long)(*token)->len))
     error_at((*token)->str, "'%s'ではありません", op);
   *token = (*token)->next;
 }
